requirements.txt
# ============================================
# SafeMolGen-DrugOracle Dependencies
# ============================================
# Core Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
torchmetrics>=1.0.0
# Chemistry & Molecules
rdkit>=2023.03.1
# Data Sources
pytdc>=0.4.1
pandas>=2.0.0
numpy>=1.24.0
# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0
# Transformers (for Generator in Phase 3)
transformers>=4.30.0
# Explainability (for Oracle in Phase 2)
shap>=0.42.0
captum>=0.6.0
# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
py3Dmol>=2.0.0
# Web App (for Phase 4)
streamlit>=1.24.0
streamlit-ketcher>=0.0.1
# Utilities
tqdm>=4.65.0
pyyaml>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0
# Experiment Tracking (optional)
wandb>=0.15.0
# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
